+++ tests/TestCases/ContainerTestCase.php
@@ -155,6 +155,7 @@ abstract class ContainerTestCase extends PHPUnit_Framework_TestCase
 		// Rocketeer
 		$config->shouldReceive('get')->with('rocketeer::default')->andReturn(array('production', 'staging'));
 		$config->shouldReceive('get')->with('rocketeer::connections')->andReturn(array());
+		$config->shouldReceive('get')->with('rocketeer::remote.strategy')->andReturn('clone');
 		$config->shouldReceive('get')->with('rocketeer::remote.application_name')->andReturn('foobar');
 		$config->shouldReceive('get')->with('rocketeer::remote.keep_releases')->andReturn(1);
 		$config->shouldReceive('get')->with('rocketeer::remote.permissions.callback')->andReturn(function ($task, $file) {